requirements.txt
amqp==5.2.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
blessed==1.20.0
bpython==0.24
cachetools==5.3.3
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
curtsies==0.4.2
cwcwidth==0.1.9
Django==4.2.13
django-celery-beat==2.6.0
django-timezone-field==6.1.0
djangorestframework==3.15.2
exif==1.6.0
google-ai-generativelanguage==0.6.5
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.0
googleapis-common-protos==1.63.1
greenlet==3.0.3
grpcio==1.64.1
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
kombu==5.3.7
pillow==10.3.0
plum-py==0.8.7
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyparsing==3.1.2
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyxdg==0.28
redis==5.0.6
requests==2.32.3
rsa==4.9
six==1.16.0
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
